% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{dark_theme}
\alias{dark_theme}
\title{Switch to dark theme.}
\usage{
dark_theme()
}
\description{
Dark theme defaults to Material.
Dark theme can be changed with \code{\link{set_dark_theme}}.
This function is executed by ThemeSwitcher addin "Dark Theme".
It throws an error if \code{getOption("themeswitcher.dark")} does not return a valid theme name.
}
