% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{light_theme}
\alias{light_theme}
\title{Switch to light theme.}
\usage{
light_theme()
}
\description{
Light theme defaults to TextMate.
Light theme can be changed with \code{\link{set_light_theme}}.
This function is executed by ThemeSwitcher addin "Light Theme".
It throws an error if \code{getOption("themeswitcher.light")} does not return a valid theme name.
}
