% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set_dark_theme}
\alias{set_dark_theme}
\title{Change dark theme for ThemeSwitcher addin}
\usage{
set_dark_theme(theme = "Material")
}
\arguments{
\item{theme}{A valid theme name as a string}
}
\value{
Sets option for themeswitcher.dark.
}
\description{
Change dark theme for ThemeSwitcher addin
}
\examples{
set_light_theme("Monokai")
}
