% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set_light_theme}
\alias{set_light_theme}
\title{Change light theme for ThemeSwitcher addin}
\usage{
set_light_theme(theme = "TextMate")
}
\arguments{
\item{theme}{A valid theme name as a string}
}
\value{
Sets option for themeswitcher.light.
}
\description{
Change light theme for ThemeSwitcher addin
}
\examples{
set_light_theme("Chrome")
}
